bot-telegram/requirements.txt
python-telegram-bot==11.1.0
requests==2.21.0
urllib3==1.24.1
validators==0.12.4